Move options utils to builtins module
