+++ boa_engine/src/builtins/intl/locale/options.rs
@@ -1,6 +1,6 @@
 use icu_locid::extensions::unicode::Value;
 
-use crate::{builtins::intl::options::OptionType, Context, JsNativeError};
+use crate::{builtins::options::OptionType, Context, JsNativeError};
 
 impl OptionType for Value {
     fn from_value(value: crate::JsValue, context: &mut Context<'_>) -> crate::JsResult<Self> {
